Improve schema and resolver usage
