+++ packages/cli/src/authentication/templates/user.schema.tpl.ts
@@ -8,7 +8,7 @@ const template = ({
   authStrategies,
   type,
   relative
-}: AuthenticationGeneratorContext) => /* ts */ `import { getValidator, getDataValidator, resolve, querySyntax } from '@feathersjs/schema'
+}: AuthenticationGeneratorContext) => /* ts */ `import { resolve, jsonSchema } from '@feathersjs/schema'
 import type { FromSchema } from '@feathersjs/schema'
 ${authStrategies.includes('local') ? `import { passwordHash } from '@feathersjs/authentication-local'` : ''}
 
@@ -35,7 +35,7 @@ export const ${camelName}DataSchema = {
 
 export type ${upperName}Data = FromSchema<typeof ${camelName}DataSchema>
 
-export const ${camelName}DataValidator = getDataValidator(${camelName}DataSchema, dataValidator)
+export const ${camelName}DataValidator = jsonSchema.getDataValidator(${camelName}DataSchema, dataValidator)
 
 export const ${camelName}DataResolver = resolve<${upperName}Data, HookContext>({
   properties: {
@@ -76,13 +76,13 @@ export const ${camelName}QuerySchema = {
   type: 'object',
   additionalProperties: false,
   properties: {
-    ...querySyntax(${camelName}Schema.properties)
+    ...jsonSchema.querySyntax(${camelName}Schema.properties)
   }
 } as const
 
 export type ${upperName}Query = FromSchema<typeof ${camelName}QuerySchema>
 
-export const ${camelName}QueryValidator = getValidator(${camelName}QuerySchema, queryValidator)
+export const ${camelName}QueryValidator = jsonSchema.getValidator(${camelName}QuerySchema, queryValidator)
 
 export const ${camelName}QueryResolver = resolve<${upperName}Query, HookContext>({
   properties: {
