+++ packages/schema/src/hooks/validate.ts
@@ -1,6 +1,7 @@
 import { HookContext, NextFunction } from '@feathersjs/feathers'
 import { BadRequest } from '@feathersjs/errors'
-import { DataValidatorMap, Schema, Validator } from '../schema'
+import { Schema, Validator } from '../schema'
+import { DataValidatorMap } from '../json-schema'
 
 export const validateQuery = <H extends HookContext>(schema: Schema<any> | Validator) => {
   const validator: Validator = typeof schema === 'function' ? schema : schema.validate.bind(schema)
